Created server utils
